+++ tests/main.test.js
@@ -1,39 +1,33 @@
 /* eslint no-console: "off" */
 
 import test from 'ava';
-import {
-    getArrayOfFaces,
-    getRandomFace,
-    prettyPrintFaces,
-    getFaceByName,
-} from '../src/textfaces.js';
-import { canonical } from '../src/faces.js';
+import txtfaces from '../src/index.js';
 
-test('should retrieve array of faces and a random face', (t) => {
-    const arrayOfFaces = getArrayOfFaces(canonical);
+test('should retrieve an array of faces and a random face', (t) => {
+    const arrayOfFaces = txtfaces.array(); 
 
     t.true(arrayOfFaces.length > 0, 'the array of faces is empty');
 
-    const face = getRandomFace(canonical);
+    const face = txtfaces.rand();
 
     t.true(arrayOfFaces.indexOf(face) >= 0, 'face not found at faces array');
 });
 
 test('should retrieve string with all faces fo pretty print', (t) => {
-    const prettyString = prettyPrintFaces(canonical);
+    const prettyString = txtfaces.prettyPrint();
     console.log(prettyString);
 
     t.assert(typeof prettyString === 'string');
 });
 
 test('should retrieve a random face by name', (t) => {
-    const face = getFaceByName(canonical, 'embarrassed');
+    const face = txtfaces.get('embarrassed');
 
     t.true([ ':$', '://)', '://3' ].indexOf(face) >= 0);
 });
 
 test('should not retrieve a random face by name because does not exists', (t) => {
-    const face = getFaceByName(canonical, 'foz');
+    const face = txtfaces.get('foz');
 
     t.is(typeof face, 'undefined');
 });
